refactor: add types.d.ts to each dir
