+++ packages/slinkity/defineConfig.d.ts
@@ -1,0 +1,3 @@
+import { UserSlinkityConfig } from './cli/types'
+
+export function defineConfig(userConfig: Partial<UserSlinkityConfig>): UserSlinkityConfig
